implmenting mocks and test of mulpliples functions
