+++ package.json
@@ -41,5 +41,8 @@
     "babel-jest": "^27.4.6",
     "jsdom": "^19.0.0",
     "lodash": "^4.17.21"
-  }
+  },
+  "jest":{
+    "testEnvironment": "jsdom"
+}
 }
